requirements.txt
django==2.0.1
djangorestframework==3.7.7
requests==2.18.4
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-api-python-client==1.6.5